Creating a covariance target, optionally by using information from KEGG pathways.

The function build.target() is a wrapper function to build the various types of covariance targets: diagonal ("D"), constant correlation ("F"), knowledge based ("G", "Gpos", and "Gstar"), correlation ("cor").

Usage

build.target(x, genegroups = NULL, type)

Arguments

x

An \(n \times p\) matrix.

genegroups

List of the groups each gene belongs to: each entry of the list is dedicated to a gene (identified the same way as in \(x\)). Each item of the list is thus a vector of pathway IDs.

type

Character string specifying the wished target: "D" for a diagonal target, "cor" for a correlation target, "G", "Gpos" and "Gstar" for a G-type target (see Jelizarow et al, 2010) and "F" for a F-target.

Value

A \(p \times p\) target covariance matrix of a certain type.

References

M. Jelizarow, V. Guillemot, A. Tenenhaus, K. Strimmer, A.-L. Boulesteix, 2010. Over-optimism in bioinformatics: an illustration. Bioinformatics. Accepted.

See also

Author

Vincent Guillemot

Examples

# Simulate dataset
x <- matrix(rnorm(20*30), 20, 30)
# Try different targets
build.target(x, type = "D")
#>           [,1]      [,2]    [,3]      [,4]     [,5]      [,6]      [,7]
#>  [1,] 1.264939 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#>  [2,] 0.000000 0.9795557 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#>  [3,] 0.000000 0.0000000 1.35106 0.0000000 0.000000 0.0000000 0.0000000
#>  [4,] 0.000000 0.0000000 0.00000 0.8342182 0.000000 0.0000000 0.0000000
#>  [5,] 0.000000 0.0000000 0.00000 0.0000000 1.027887 0.0000000 0.0000000
#>  [6,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.9506503 0.0000000
#>  [7,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.7077416
#>  [8,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#>  [9,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [10,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [11,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [12,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [13,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [14,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [15,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [16,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [17,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [18,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [19,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [20,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [21,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [22,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [23,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [24,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [25,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [26,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [27,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [28,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [29,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#> [30,] 0.000000 0.0000000 0.00000 0.0000000 0.000000 0.0000000 0.0000000
#>            [,8]     [,9]    [,10]    [,11]     [,12]     [,13]    [,14]
#>  [1,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [2,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [3,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [4,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [5,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [6,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [7,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [8,] 0.8982466 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>  [9,] 0.0000000 1.167925 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [10,] 0.0000000 0.000000 1.333264 0.000000 0.0000000 0.0000000 0.000000
#> [11,] 0.0000000 0.000000 0.000000 1.458403 0.0000000 0.0000000 0.000000
#> [12,] 0.0000000 0.000000 0.000000 0.000000 0.8168753 0.0000000 0.000000
#> [13,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.9405621 0.000000
#> [14,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 1.373228
#> [15,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [16,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [17,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [18,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [19,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [20,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [21,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [22,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [23,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [24,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [25,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [26,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [27,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [28,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [29,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#> [30,] 0.0000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000
#>          [,15]    [,16]    [,17]    [,18]     [,19]     [,20]    [,21]    [,22]
#>  [1,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [2,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [3,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [4,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [5,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [6,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [7,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [8,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>  [9,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [10,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [11,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [12,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [13,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [14,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [15,] 1.039516 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [16,] 0.000000 1.177066 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [17,] 0.000000 0.000000 1.135575 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [18,] 0.000000 0.000000 0.000000 1.180119 0.0000000 0.0000000 0.000000 0.000000
#> [19,] 0.000000 0.000000 0.000000 0.000000 0.8713316 0.0000000 0.000000 0.000000
#> [20,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.7099573 0.000000 0.000000
#> [21,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 1.292137 0.000000
#> [22,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 1.138256
#> [23,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [24,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [25,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [26,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [27,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [28,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [29,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#> [30,] 0.000000 0.000000 0.000000 0.000000 0.0000000 0.0000000 0.000000 0.000000
#>          [,23]     [,24]    [,25]     [,26]     [,27]     [,28]    [,29]
#>  [1,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [2,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [3,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [4,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [5,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [6,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [7,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [8,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>  [9,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [10,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [11,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [12,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [13,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [14,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [15,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [16,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [17,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [18,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [19,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [20,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [21,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [22,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [23,] 1.109673 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [24,] 0.000000 0.8698377 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#> [25,] 0.000000 0.0000000 0.998272 0.0000000 0.0000000 0.0000000 0.000000
#> [26,] 0.000000 0.0000000 0.000000 0.9658397 0.0000000 0.0000000 0.000000
#> [27,] 0.000000 0.0000000 0.000000 0.0000000 0.7884315 0.0000000 0.000000
#> [28,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.8452144 0.000000
#> [29,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 1.110784
#> [30,] 0.000000 0.0000000 0.000000 0.0000000 0.0000000 0.0000000 0.000000
#>          [,30]
#>  [1,] 0.000000
#>  [2,] 0.000000
#>  [3,] 0.000000
#>  [4,] 0.000000
#>  [5,] 0.000000
#>  [6,] 0.000000
#>  [7,] 0.000000
#>  [8,] 0.000000
#>  [9,] 0.000000
#> [10,] 0.000000
#> [11,] 0.000000
#> [12,] 0.000000
#> [13,] 0.000000
#> [14,] 0.000000
#> [15,] 0.000000
#> [16,] 0.000000
#> [17,] 0.000000
#> [18,] 0.000000
#> [19,] 0.000000
#> [20,] 0.000000
#> [21,] 0.000000
#> [22,] 0.000000
#> [23,] 0.000000
#> [24,] 0.000000
#> [25,] 0.000000
#> [26,] 0.000000
#> [27,] 0.000000
#> [28,] 0.000000
#> [29,] 0.000000
#> [30,] 0.983945