Shrinkage estimator of the covariance matrix, given a data set and a covariance target.

The estimator combines the sample covariance matrix \(S\) with a target matrix \(T\): $$\widehat{\Sigma} = \lambda T + (1 - \lambda)S$$ The estimated intensity \(\lambda\) is constrained to \([0, 1]\). Values near zero keep more of the sample covariance; values near one give more weight to the target.

Usage

shrink.estim(x, target)

Arguments

x

A \(n \times p\) matrix (the data set) .

target

A \(p \times p\) matrix (the covariance target).

Value

A list with components shrink.cov, the estimated covariance matrix, and lambda, the estimated shrinkage intensity.

References

J. Schaefer and K. Strimmer, 2005. A shrinkage approach to large-scale covariance matrix estimation and implications for functional genomics. Statist. Appl. Genet. Mol. Biol. 4:32. Jelizarow, M., Guillemot, V., Tenenhaus, A., Strimmer, K. and Boulesteix, A.-L. (2010). Over-optimism in bioinformatics: an illustration. Bioinformatics.

See also

Author

Monika Jelizarow and Vincent Guillemot

Examples

# Simulate dataset
x <- matrix(rnorm(20*30),20,30)
# Try different targets
shrink.estim(x, target = build.target(x, type = "D"))
#> $shrink.cov
#>                [,1]          [,2]         [,3]          [,4]          [,5]
#>  [1,]  1.139330e+00  0.0118226569 -0.002905975  0.0166263486 -0.0236555907
#>  [2,]  1.182266e-02  0.8825992585  0.006929554 -0.0032605563 -0.0119300017
#>  [3,] -2.905975e-03  0.0069295542  0.442275993 -0.0042143236  0.0053917744
#>  [4,]  1.662635e-02 -0.0032605563 -0.004214324  1.0955377929 -0.0087081072
#>  [5,] -2.365559e-02 -0.0119300017  0.005391774 -0.0087081072  1.4443585763
#>  [6,]  5.541830e-03  0.0068125296 -0.007389681  0.0061700377 -0.0151609102
#>  [7,]  2.030756e-02  0.0197300026 -0.004073800  0.0058609931 -0.0132811658
#>  [8,]  1.268787e-02  0.0067921986 -0.001222036  0.0015888480  0.0007283133
#>  [9,] -6.606071e-03 -0.0019237442  0.008046988  0.0053508284  0.0069176771
#> [10,] -1.108331e-02 -0.0151810503  0.001080459 -0.0043205704  0.0015244476
#> [11,] -1.584156e-02  0.0041105239  0.008052731  0.0011236760 -0.0016978189
#> [12,]  9.408474e-03  0.0099766017  0.018574410 -0.0084493639 -0.0123032835
#> [13,] -3.116515e-02 -0.0206964904 -0.013694948 -0.0081041483  0.0100918338
#> [14,] -3.043136e-03 -0.0002753102 -0.006365771  0.0025732239 -0.0112318357
#> [15,]  7.908769e-05  0.0149675141 -0.005557176 -0.0085796046 -0.0061583611
#> [16,]  5.630158e-03  0.0066237270  0.000947962  0.0074569289 -0.0192482087
#> [17,]  2.851304e-03  0.0018797599  0.010806296 -0.0003207684 -0.0069793212
#> [18,] -1.030878e-02 -0.0157950347 -0.003722096 -0.0085535034  0.0088262601
#> [19,]  6.728189e-04  0.0187710577 -0.002317695  0.0190644294 -0.0093682436
#> [20,]  4.685824e-03 -0.0109371704  0.005646009  0.0165963045 -0.0112028092
#> [21,] -2.858424e-02  0.0008518814 -0.006517552  0.0135294360 -0.0004127660
#> [22,] -2.038744e-03 -0.0079833173  0.008295906 -0.0158092461  0.0191293263
#> [23,]  2.443592e-03 -0.0010810604 -0.007059648 -0.0041744971  0.0063985248
#> [24,] -6.341096e-03  0.0024995190 -0.006605725  0.0027496378 -0.0262387953
#> [25,] -1.214108e-02 -0.0081043301  0.009912442  0.0097294724 -0.0057464802
#> [26,] -4.675842e-03 -0.0011555778 -0.006830360  0.0043683661 -0.0051352710
#> [27,]  8.453024e-03 -0.0045287282  0.001032371  0.0087494354 -0.0015208560
#> [28,] -3.117251e-03 -0.0060482783  0.009151263  0.0166799753  0.0124351820
#> [29,]  1.587528e-02  0.0028484949 -0.007243732  0.0231309089 -0.0118431232
#> [30,] -7.767582e-03 -0.0128959788 -0.010340418 -0.0084187943  0.0186968326
#>                [,6]          [,7]          [,8]         [,9]         [,10]
#>  [1,]  5.541830e-03  0.0203075599  0.0126878697 -0.006606071 -0.0110833115
#>  [2,]  6.812530e-03  0.0197300026  0.0067921986 -0.001923744 -0.0151810503
#>  [3,] -7.389681e-03 -0.0040738000 -0.0012220356  0.008046988  0.0010804591
#>  [4,]  6.170038e-03  0.0058609931  0.0015888480  0.005350828 -0.0043205704
#>  [5,] -1.516091e-02 -0.0132811658  0.0007283133  0.006917677  0.0015244476
#>  [6,]  1.204240e+00  0.0143958367  0.0022443191 -0.013898851 -0.0013729536
#>  [7,]  1.439584e-02  0.8478877487  0.0087559963 -0.012295356  0.0028794766
#>  [8,]  2.244319e-03  0.0087559963  0.5779814818 -0.011203441  0.0087232661
#>  [9,] -1.389885e-02 -0.0122953558 -0.0112034407  0.770292953 -0.0204705450
#> [10,] -1.372954e-03  0.0028794766  0.0087232661 -0.020470545  1.0212897930
#> [11,]  1.385571e-02 -0.0011115416  0.0142243681 -0.011411840  0.0064219208
#> [12,]  1.142075e-02 -0.0114762795 -0.0013479999 -0.001139531 -0.0088229816
#> [13,] -3.789308e-03 -0.0189932617 -0.0252530214  0.012676859 -0.0103588485
#> [14,]  1.551991e-02  0.0034900014  0.0046026862 -0.013565534  0.0157880116
#> [15,]  1.068673e-02  0.0003587386  0.0022608103 -0.002833532 -0.0161611803
#> [16,]  9.617159e-03 -0.0019538318  0.0012806598  0.010319364 -0.0170646378
#> [17,] -5.001626e-03 -0.0011598474  0.0077948832 -0.005723314 -0.0007637756
#> [18,] -2.037380e-02 -0.0042927519  0.0003877171  0.007998708  0.0182523622
#> [19,]  4.988391e-03 -0.0085656964 -0.0034168535  0.002236013 -0.0114336455
#> [20,]  9.731546e-03 -0.0018005966 -0.0169937195 -0.002731459  0.0052340781
#> [21,] -1.854162e-03  0.0029339138 -0.0052886741  0.018922160  0.0005438913
#> [22,] -1.935353e-02 -0.0028374767  0.0088503624 -0.017185875  0.0297698632
#> [23,] -4.436013e-03 -0.0003370341 -0.0048277208 -0.003870319  0.0007128951
#> [24,]  4.384986e-02 -0.0037441718  0.0043763992 -0.016423402 -0.0041032622
#> [25,] -3.136428e-02 -0.0196739036 -0.0006970691 -0.004811411  0.0108789301
#> [26,]  1.115233e-02 -0.0090446184  0.0043952996 -0.011884842  0.0071130962
#> [27,] -1.579305e-03 -0.0210061124  0.0024745573  0.016081054 -0.0041005926
#> [28,] -3.203011e-02 -0.0046432286 -0.0008839866  0.014221911 -0.0105344461
#> [29,]  3.902280e-03  0.0187275878  0.0053521641 -0.011319104 -0.0088217529
#> [30,] -5.139774e-06 -0.0020545923 -0.0109536388 -0.014791841  0.0018586401
#>               [,11]         [,12]         [,13]         [,14]         [,15]
#>  [1,] -0.0158415607  0.0094084738 -3.116515e-02 -0.0030431358  7.908769e-05
#>  [2,]  0.0041105239  0.0099766017 -2.069649e-02 -0.0002753102  1.496751e-02
#>  [3,]  0.0080527305  0.0185744100 -1.369495e-02 -0.0063657706 -5.557176e-03
#>  [4,]  0.0011236760 -0.0084493639 -8.104148e-03  0.0025732239 -8.579605e-03
#>  [5,] -0.0016978189 -0.0123032835  1.009183e-02 -0.0112318357 -6.158361e-03
#>  [6,]  0.0138557080  0.0114207515 -3.789308e-03  0.0155199074  1.068673e-02
#>  [7,] -0.0011115416 -0.0114762795 -1.899326e-02  0.0034900014  3.587386e-04
#>  [8,]  0.0142243681 -0.0013479999 -2.525302e-02  0.0046026862  2.260810e-03
#>  [9,] -0.0114118402 -0.0011395308  1.267686e-02 -0.0135655338 -2.833532e-03
#> [10,]  0.0064219208 -0.0088229816 -1.035885e-02  0.0157880116 -1.616118e-02
#> [11,]  0.9708617634  0.0151670456 -8.181231e-03  0.0049564927  1.296226e-02
#> [12,]  0.0151670456  1.3987028190 -1.927434e-02 -0.0011938740  7.826071e-03
#> [13,] -0.0081812312 -0.0192743422  1.105734e+00 -0.0021256299 -7.828255e-04
#> [14,]  0.0049564927 -0.0011938740 -2.125630e-03  0.5207121400  3.594574e-03
#> [15,]  0.0129622617  0.0078260706 -7.828255e-04  0.0035945743  7.344532e-01
#> [16,]  0.0055110796  0.0114640029  2.129342e-03  0.0022668413  9.484622e-03
#> [17,]  0.0130133852  0.0213019461 -1.609370e-02 -0.0148653012  2.611751e-04
#> [18,] -0.0124920354 -0.0186476624  3.443572e-03 -0.0018758882 -1.041372e-02
#> [19,] -0.0002915603 -0.0014585414  4.446365e-04 -0.0061347942  1.302737e-02
#> [20,] -0.0082242803  0.0095082255  1.188015e-05  0.0020478632 -1.753996e-02
#> [21,]  0.0097808624 -0.0212263490  1.594889e-02 -0.0034449602  6.918831e-03
#> [22,] -0.0038396358  0.0028684216 -1.379023e-02  0.0032520821 -9.414387e-03
#> [23,] -0.0097702311 -0.0015784995  1.606469e-03  0.0009186465  9.124360e-03
#> [24,]  0.0314891995  0.0212893078 -2.990366e-03  0.0129366371  2.355464e-02
#> [25,]  0.0173539321  0.0129419909 -4.474952e-03 -0.0068237672  4.342135e-03
#> [26,]  0.0037605520 -0.0077134679  4.049469e-03  0.0042121448 -4.956913e-03
#> [27,]  0.0008580644  0.0211036374  1.131753e-04 -0.0047631025  6.902464e-03
#> [28,]  0.0014357962  0.0004330408  2.357919e-04 -0.0282405709 -1.409012e-02
#> [29,]  0.0050840484 -0.0146600054  3.139340e-03  0.0016637640 -3.506402e-03
#> [30,] -0.0065581634  0.0010635191  1.758592e-02 -0.0049389551  2.276969e-03
#>              [,16]         [,17]         [,18]         [,19]         [,20]
#>  [1,]  0.005630158  2.851304e-03 -0.0103087791  0.0006728189  4.685824e-03
#>  [2,]  0.006623727  1.879760e-03 -0.0157950347  0.0187710577 -1.093717e-02
#>  [3,]  0.000947962  1.080630e-02 -0.0037220958 -0.0023176954  5.646009e-03
#>  [4,]  0.007456929 -3.207684e-04 -0.0085535034  0.0190644294  1.659630e-02
#>  [5,] -0.019248209 -6.979321e-03  0.0088262601 -0.0093682436 -1.120281e-02
#>  [6,]  0.009617159 -5.001626e-03 -0.0203737963  0.0049883913  9.731546e-03
#>  [7,] -0.001953832 -1.159847e-03 -0.0042927519 -0.0085656964 -1.800597e-03
#>  [8,]  0.001280660  7.794883e-03  0.0003877171 -0.0034168535 -1.699372e-02
#>  [9,]  0.010319364 -5.723314e-03  0.0079987084  0.0022360128 -2.731459e-03
#> [10,] -0.017064638 -7.637756e-04  0.0182523622 -0.0114336455  5.234078e-03
#> [11,]  0.005511080  1.301339e-02 -0.0124920354 -0.0002915603 -8.224280e-03
#> [12,]  0.011464003  2.130195e-02 -0.0186476624 -0.0014585414  9.508225e-03
#> [13,]  0.002129342 -1.609370e-02  0.0034435721  0.0004446365  1.188015e-05
#> [14,]  0.002266841 -1.486530e-02 -0.0018758882 -0.0061347942  2.047863e-03
#> [15,]  0.009484622  2.611751e-04 -0.0104137224  0.0130273695 -1.753996e-02
#> [16,]  0.608683056 -7.893644e-03 -0.0120116281  0.0056305988 -4.496375e-03
#> [17,] -0.007893644  1.236254e+00 -0.0132364459  0.0133099734  1.916055e-03
#> [18,] -0.012011628 -1.323645e-02  0.6586447775 -0.0153373129 -4.359088e-03
#> [19,]  0.005630599  1.330997e-02 -0.0153373129  1.2675541840 -9.115890e-03
#> [20,] -0.004496375  1.916055e-03 -0.0043590880 -0.0091158896  7.716244e-01
#> [21,]  0.002838472 -3.476379e-05  0.0105312274  0.0196752558 -1.278302e-02
#> [22,] -0.013430494  1.413659e-02  0.0107249036 -0.0308601413  9.096214e-03
#> [23,] -0.003552502 -9.724143e-03  0.0077340343 -0.0070850533  2.114982e-03
#> [24,]  0.012886881  3.163498e-03 -0.0187923250  0.0180598346  2.521229e-04
#> [25,] -0.004901546  2.630505e-02  0.0050742264  0.0214414907 -4.426547e-04
#> [26,] -0.004163717  5.177122e-04 -0.0042170591  0.0147171636  4.553688e-03
#> [27,]  0.001828944  1.146757e-02  0.0034092988  0.0001735026  4.069156e-03
#> [28,]  0.004718473  2.209383e-02  0.0042281642  0.0028522372  3.413351e-03
#> [29,]  0.010022143  1.008315e-02 -0.0220082641 -0.0021626999  5.913150e-03
#> [30,] -0.012917283 -3.063921e-04  0.0001661769  0.0078797871 -2.607739e-03
#>               [,21]        [,22]         [,23]         [,24]         [,25]
#>  [1,] -2.858424e-02 -0.002038744  0.0024435917 -0.0063410960 -0.0121410828
#>  [2,]  8.518814e-04 -0.007983317 -0.0010810604  0.0024995190 -0.0081043301
#>  [3,] -6.517552e-03  0.008295906 -0.0070596483 -0.0066057250  0.0099124424
#>  [4,]  1.352944e-02 -0.015809246 -0.0041744971  0.0027496378  0.0097294724
#>  [5,] -4.127660e-04  0.019129326  0.0063985248 -0.0262387953 -0.0057464802
#>  [6,] -1.854162e-03 -0.019353529 -0.0044360131  0.0438498644 -0.0313642768
#>  [7,]  2.933914e-03 -0.002837477 -0.0003370341 -0.0037441718 -0.0196739036
#>  [8,] -5.288674e-03  0.008850362 -0.0048277208  0.0043763992 -0.0006970691
#>  [9,]  1.892216e-02 -0.017185875 -0.0038703189 -0.0164234020 -0.0048114110
#> [10,]  5.438913e-04  0.029769863  0.0007128951 -0.0041032622  0.0108789301
#> [11,]  9.780862e-03 -0.003839636 -0.0097702311  0.0314891995  0.0173539321
#> [12,] -2.122635e-02  0.002868422 -0.0015784995  0.0212893078  0.0129419909
#> [13,]  1.594889e-02 -0.013790225  0.0016064691 -0.0029903657 -0.0044749523
#> [14,] -3.444960e-03  0.003252082  0.0009186465  0.0129366371 -0.0068237672
#> [15,]  6.918831e-03 -0.009414387  0.0091243599  0.0235546353  0.0043421348
#> [16,]  2.838472e-03 -0.013430494 -0.0035525022  0.0128868808 -0.0049015456
#> [17,] -3.476379e-05  0.014136587 -0.0097241428  0.0031634976  0.0263050488
#> [18,]  1.053123e-02  0.010724904  0.0077340343 -0.0187923250  0.0050742264
#> [19,]  1.967526e-02 -0.030860141 -0.0070850533  0.0180598346  0.0214414907
#> [20,] -1.278302e-02  0.009096214  0.0021149818  0.0002521229 -0.0004426547
#> [21,]  1.116174e+00 -0.028634019 -0.0048135766  0.0083375366  0.0078848428
#> [22,] -2.863402e-02  1.519084059  0.0181789881 -0.0343261446  0.0091705255
#> [23,] -4.813577e-03  0.018178988  0.4170620751 -0.0098166521 -0.0015597259
#> [24,]  8.337537e-03 -0.034326145 -0.0098166521  1.3782356739  0.0045156304
#> [25,]  7.884843e-03  0.009170526 -0.0015597259  0.0045156304  1.2941083687
#> [26,] -2.583499e-03 -0.003554904 -0.0075972117  0.0193335766  0.0022148795
#> [27,] -2.885268e-03  0.017625194  0.0082351541 -0.0031365463  0.0085303011
#> [28,]  1.086923e-02  0.011874693 -0.0007850889 -0.0239627999  0.0272088583
#> [29,] -5.328310e-03  0.004843508 -0.0063631160 -0.0042429387 -0.0046088494
#> [30,] -2.443477e-03 -0.007054446  0.0050311551  0.0070050371  0.0097961487
#>               [,26]         [,27]         [,28]        [,29]         [,30]
#>  [1,] -0.0046758415  0.0084530240 -0.0031172509  0.015875281 -7.767582e-03
#>  [2,] -0.0011555778 -0.0045287282 -0.0060482783  0.002848495 -1.289598e-02
#>  [3,] -0.0068303597  0.0010323714  0.0091512625 -0.007243732 -1.034042e-02
#>  [4,]  0.0043683661  0.0087494354  0.0166799753  0.023130909 -8.418794e-03
#>  [5,] -0.0051352710 -0.0015208560  0.0124351820 -0.011843123  1.869683e-02
#>  [6,]  0.0111523273 -0.0015793051 -0.0320301054  0.003902280 -5.139774e-06
#>  [7,] -0.0090446184 -0.0210061124 -0.0046432286  0.018727588 -2.054592e-03
#>  [8,]  0.0043952996  0.0024745573 -0.0008839866  0.005352164 -1.095364e-02
#>  [9,] -0.0118848423  0.0160810540  0.0142219113 -0.011319104 -1.479184e-02
#> [10,]  0.0071130962 -0.0041005926 -0.0105344461 -0.008821753  1.858640e-03
#> [11,]  0.0037605520  0.0008580644  0.0014357962  0.005084048 -6.558163e-03
#> [12,] -0.0077134679  0.0211036374  0.0004330408 -0.014660005  1.063519e-03
#> [13,]  0.0040494691  0.0001131753  0.0002357919  0.003139340  1.758592e-02
#> [14,]  0.0042121448 -0.0047631025 -0.0282405709  0.001663764 -4.938955e-03
#> [15,] -0.0049569127  0.0069024641 -0.0140901222 -0.003506402  2.276969e-03
#> [16,] -0.0041637168  0.0018289442  0.0047184729  0.010022143 -1.291728e-02
#> [17,]  0.0005177122  0.0114675716  0.0220938334  0.010083154 -3.063921e-04
#> [18,] -0.0042170591  0.0034092988  0.0042281642 -0.022008264  1.661769e-04
#> [19,]  0.0147171636  0.0001735026  0.0028522372 -0.002162700  7.879787e-03
#> [20,]  0.0045536882  0.0040691561  0.0034133511  0.005913150 -2.607739e-03
#> [21,] -0.0025834995 -0.0028852684  0.0108692304 -0.005328310 -2.443477e-03
#> [22,] -0.0035549036  0.0176251939  0.0118746930  0.004843508 -7.054446e-03
#> [23,] -0.0075972117  0.0082351541 -0.0007850889 -0.006363116  5.031155e-03
#> [24,]  0.0193335766 -0.0031365463 -0.0239627999 -0.004242939  7.005037e-03
#> [25,]  0.0022148795  0.0085303011  0.0272088583 -0.004608849  9.796149e-03
#> [26,]  0.5915493244  0.0051147786 -0.0062719391  0.003666054  2.069697e-03
#> [27,]  0.0051147786  1.2825809893  0.0053291083 -0.008777526 -1.655243e-02
#> [28,] -0.0062719391  0.0053291083  1.1855461505  0.014472274  7.125653e-04
#> [29,]  0.0036660539 -0.0087775259  0.0144722740  1.011276138 -3.691577e-03
#> [30,]  0.0020696972 -0.0165524253  0.0007125653 -0.003691577  8.660601e-01
#> 
#> $lambda
#> [1] 0.9500492
#>