Significance-filtered knowledge-based target.

The \(p \times p\) target Cor is computed from the \(n \times p\) data matrix. It it a modified version of target G. In particular, it tests the correlations (with a significance level of 0.05) and sets the non-significant correlations to zero before the mean correlation \(\bar{r}\) is computed.

Usage

targetCor(x, genegroups)

Arguments

x

A \(n \times p\) data matrix.

genegroups

A list of genes obtained using the database KEGG, where each entry itself is a list of pathway names this genes belongs to. If a gene does not belong to any gene functional group, the entry is NA.

Value

A \(p \times p\) matrix.

Details

For retained linked pairs, the off-diagonal target is $$T_{ij} = \bar{r}_{sig}\sqrt{s_{ii}s_{jj}}$$ where \(\bar{r}_{sig}\) is the mean correlation among links whose correlation test p-value is below 0.05.

See also

Author

Monika Jelizarow and Vincent Guillemot

Examples

# A short example on a toy dataset
# require(SHIP)
data(expl)
target_matrix <- targetCor(expl$x, expl$genegroups)
which(target_matrix[upper.tri(target_matrix)] != 0) # not many non zero coefficients !
#>    [1]    1    2    3    7    8    9   16   17   18   20   29   30   31   33
#>   [15]   35   37   38   39   41   43   45   56   57   58   60   62   64   65
#>   [29]   67   68   69   71   73   75   76   78   79   80   81   83   85   87
#>   [43]   88   90   91   92   93   94   96   98  100  101  103  104  105  106
#>   [57]  107  108  110  112  114  115  117  118  119  120  121  122  123  125
#>   [71]  127  129  130  132  133  134  135  136  137  138  139  141  143  145
#>   [85]  146  148  149  150  151  152  153  154  155  156  158  160  162  163
#>   [99]  165  166  167  168  169  170  171  172  173  174  176  178  180  181
#>  [113]  183  184  185  186  187  188  189  190  211  212  213  215  217  219
#>  [127]  220  222  223  224  225  226  227  228  229  230  232  233  234  236
#>  [141]  238  240  241  243  244  245  246  247  248  249  250  251  253  277
#>  [155]  278  279  281  283  285  286  288  289  290  291  292  293  294  295
#>  [169]  296  298  299  301  302  303  305  307  309  310  312  313  314  315
#>  [183]  316  317  318  319  320  322  323  325  326  327  328  330  332  334
#>  [197]  335  337  338  339  340  341  342  343  344  345  347  348  350  351
#>  [211]  379  380  381  383  385  387  388  390  391  392  393  394  395  396
#>  [225]  397  398  400  401  403  404  405  407  408  409  411  413  415  416
#>  [239]  418  419  420  421  422  423  424  425  426  428  429  431  432  433
#>  [253]  435  529  530  531  533  535  537  538  540  541  542  543  544  545
#>  [267]  546  547  548  551  553  554  555  557  558  562  563  564  566  568
#>  [281]  571  575  576  577  578  579  580  581  584  586  587  588  590  591
#>  [295]  595  639  665  667  668  669  671  673  675  676  678  679  680  681
#>  [309]  682  683  684  685  686  688  689  691  692  693  695  696  700  701
#>  [323]  704  705  706  708  710  712  713  715  716  717  718  719  720  721
#>  [337]  722  723  725  726  728  729  730  732  733  737  738  741  788  931
#>  [351]  947  948  949  951  953  955  956  958  959  960  961  962  963  964
#>  [365]  965  966  968  969  971  972  973  975  976  980  981  984  985 1082
#>  [379] 1083 1084 1086 1088 1090 1091 1093 1094 1095 1096 1097 1098 1099 1100
#>  [393] 1101 1103 1104 1106 1107 1108 1110 1111 1115 1116 1119 1120 1126 1129
#>  [407] 1133 1140 1142 1143 1145 1146 1147 1154 1157 1158 1162 1163 1166 1167
#>  [421] 1173 1176 1177 1178 1179 1181 1183 1185 1186 1188 1189 1190 1191 1192
#>  [435] 1193 1194 1195 1196 1199 1201 1202 1203 1205 1206 1210 1211 1213 1214
#>  [449] 1215 1221 1224 1226 1227 1228 1230 1232 1234 1235 1237 1238 1239 1240
#>  [463] 1241 1242 1243 1244 1245 1247 1248 1250 1251 1252 1254 1255 1259 1263
#>  [477] 1264 1270 1273 1274 1275 1277 1287 1288 1290 1292 1293 1294 1305 1312
#>  [491] 1323 1325 1335 1338 1339 1341 1363 1376 1378 1379 1380 1381 1383 1385
#>  [505] 1387 1388 1390 1391 1392 1393 1394 1395 1396 1397 1398 1400 1401 1403
#>  [519] 1404 1405 1407 1408 1412 1413 1416 1417 1423 1426 1427 1428 1429 1486
#>  [533] 1487 1488 1490 1492 1494 1495 1497 1498 1499 1500 1501 1502 1503 1504
#>  [547] 1505 1507 1508 1510 1511 1512 1514 1515 1519 1520 1523 1524 1530 1533
#>  [561] 1535 1536 1539 1541 1542 1543 1547 1550 1552 1554 1555 1557 1558 1560
#>  [575] 1563 1565 1567 1569 1570 1575 1578 1579 1585 1589 1592 1593 1594 1596
#>  [589] 1597 1598 1599 1601 1603 1606 1608 1609 1610 1611 1612 1613 1614 1615
#>  [603] 1616 1618 1619 1621 1622 1623 1625 1626 1634 1635 1641 1644 1649 1650
#>  [617] 1652 1653 1654 1655 1656 1658 1660 1663 1665 1666 1667 1669 1670 1671
#>  [631] 1672 1673 1675 1676 1678 1679 1680 1682 1683 1691 1692 1698 1701 1705
#>  [645] 1707 1709 1711 1745 1746 1748 1771 1772 1773 1775 1777 1779 1780 1782
#>  [659] 1783 1784 1785 1786 1787 1788 1789 1790 1792 1793 1795 1796 1797 1799
#>  [673] 1800 1804 1805 1808 1809 1815 1818 1819 1820 1821 1824 1826 1827 1828
#>  [687] 1829 1993 2040 2081 2082 2083 2085 2087 2089 2090 2092 2093 2094 2095
#>  [701] 2096 2097 2098 2099 2100 2102 2103 2105 2106 2107 2109 2110 2115 2118
#>  [715] 2119 2125 2128 2131 2133 2134 2136 2138 2139 2141 2160 2161 2165 2194
#>  [729] 2197 2198 2202 2226 2229 2246 2248 2260 2261 2263 2264 2268 2269 2278
#>  [743] 2416 2417 2418 2420 2422 2424 2425 2427 2428 2429 2430 2431 2432 2433
#>  [757] 2434 2435 2437 2438 2440 2441 2442 2444 2445 2453 2454 2460 2463 2465
#>  [771] 2466 2469 2471 2472 2473 2474 2476 2481 2482 2563 2571 2572 2576 2579
#>  [785] 2590 2591 2605 2608 2612 2614 2623 2627 2629 2630 2631 2633 2635 2637
#>  [799] 2638 2640 2642 2643 2644 2645 2646 2647 2648 2651 2653 2654 2655 2657
#>  [813] 2658 2665 2666 2667 2673 2676 2679 2681 2682 2684 2686 2687 2689 2694
#>  [827] 2776 2777 2778 2780 2782 2784 2785 2787 2788 2789 2790 2791 2792 2793
#>  [841] 2794 2795 2797 2798 2800 2801 2802 2804 2805 2809 2810 2813 2814 2820
#>  [855] 2823 2824 2825 2826 2829 2831 2832 2833 2834 2836 2841 2846 2849 2975
#>  [869] 2977 3004 3005 3006 3008 3010 3012 3013 3015 3016 3017 3018 3019 3020
#>  [883] 3021 3022 3023 3025 3026 3028 3029 3030 3032 3033 3037 3041 3042 3048
#>  [897] 3051 3053 3054 3057 3059 3060 3061 3062 3064 3069 3074 3076 3077 3079
#>  [911] 3241 3242 3243 3245 3247 3249 3250 3252 3253 3254 3255 3256 3257 3258
#>  [925] 3259 3260 3262 3263 3265 3266 3267 3269 3270 3274 3275 3278 3279 3285
#>  [939] 3288 3289 3290 3291 3294 3296 3297 3298 3299 3301 3306 3311 3314 3316
#>  [953] 3319 3322 3323 3324 3326 3328 3330 3331 3333 3334 3335 3336 3337 3338
#>  [967] 3339 3340 3341 3343 3344 3346 3347 3348 3350 3351 3355 3356 3359 3360
#>  [981] 3366 3369 3370 3371 3372 3373 3375 3377 3378 3379 3380 3382 3387 3392
#>  [995] 3395 3397 3400 3403 3475 3487 3488 3489 3491 3493 3495 3496 3498 3499
#> [1009] 3500 3501 3502 3503 3504 3505 3506 3508 3509 3511 3512 3513 3515 3516
#> [1023] 3520 3521 3524 3525 3531 3534 3535 3536 3537 3540 3542 3543 3544 3545
#> [1037] 3547 3552 3557 3560 3562 3565 3568 3569 3633 3656 3657 3658 3660 3662
#> [1051] 3664 3665 3667 3668 3669 3670 3671 3672 3673 3674 3675 3678 3680 3681
#> [1065] 3682 3684 3685 3690 3692 3693 3694 3700 3703 3704 3705 3707 3708 3709
#> [1079] 3711 3712 3715 3716 3721 3722 3723 3726 3731 3734 3737 3738 3740 3742
#> [1093] 3743 3744 3746 3748 3750 3751 3753 3754 3755 3756 3757 3758 3759 3760
#> [1107] 3761 3763 3764 3766 3767 3768 3770 3771 3775 3776 3779 3780 3786 3789
#> [1121] 3791 3792 3795 3797 3798 3799 3800 3802 3807 3812 3815 3817 3820 3823
#> [1135] 3824 3826 3828 3829 3830 3831 3833 3835 3837 3838 3840 3841 3842 3843
#> [1149] 3844 3845 3846 3847 3848 3850 3851 3853 3854 3855 3857 3858 3862 3863
#> [1163] 3866 3867 3873 3876 3878 3879 3882 3884 3885 3886 3887 3889 3894 3895
#> [1177] 3899 3902 3904 3907 3910 3911 3913 3915 3916 3917 3918 3919 3921 3923
#> [1191] 3925 3926 3928 3929 3930 3931 3932 3933 3934 3935 3936 3938 3939 3941
#> [1205] 3942 3943 3945 3946 3950 3951 3954 3955 3961 3964 3965 3966 3967 3970
#> [1219] 3972 3973 3975 3977 3987 3992 3995 3998 3999 4001 4003 4004 4005 4006
#> [1233] 4007 4008 4010 4012 4014 4015 4017 4018 4019 4020 4021 4022 4023 4024
#> [1247] 4025 4027 4028 4030 4031 4032 4034 4035 4039 4040 4043 4044 4050 4053
#> [1261] 4054 4055 4056 4059 4061 4062 4063 4064 4066 4071 4076 4079 4081 4084
#> [1275] 4087 4088 4090 4092 4093 4094 4095 4372 4373 4374 4376 4378 4380 4381
#> [1289] 4383 4384 4385 4386 4387 4388 4389 4390 4391 4393 4394 4396 4397 4398
#> [1303] 4400 4401 4409 4410 4416 4419 4422 4425 4427 4429 4430 4432 4437 4438
#> [1317] 4442 4444 4445 4447 4450 4453 4454 4456 4459 4460 4461 4462 4496 4559
#> [1331] 4562 4563 4565 4567 4570 4572 4574 4575 4577 4586 4587 4589 4590 4594
#> [1345] 4598 4599 4605 4609 4610 4611 4614 4616 4618 4621 4631 4638 4642 4645
#> [1359] 4649 4651 4689 4754 4755 4756 4758 4760 4762 4763 4765 4766 4767 4768
#> [1373] 4769 4770 4771 4772 4773 4775 4776 4778 4779 4780 4782 4783 4787 4788
#> [1387] 4791 4792 4798 4801 4803 4804 4806 4807 4809 4810 4811 4814 4819 4824
#> [1401] 4829 4832 4835 4836 4838 4840 4841 4842 4843 4844 4848 4852 4853 4854
#> [1415] 4856 4858 4860 4863 4864 4865 4866 4867 4868 4870 4876 4877 4878 4880
#> [1429] 4881 4885 4886 4889 4890 4896 4899 4900 4901 4902 4905 4907 4911 4912
#> [1443] 4917 4922 4924 4927 4929 4933 4934 4936 4938 4939 4940 4941 4942 4948