Positive-correlation knowledge-based target.

This is a version of targetG() that estimates the common correlation from positive linked correlations only. Negative linked correlations contribute zero to the off-diagonal target. $$T_{ij} = \bar{r}_+\sqrt{s_{ii}s_{jj}}$$ for linked pairs with positive correlation, where \(\bar{r}_+\) is their mean positive correlation.

Usage

targetGpos(x, genegroups)

Arguments

x

A \(n \times p\) data matrix.

genegroups

A list of genes obtained using the database KEGG, where each entry itself is a list of pathway names this genes belongs to. If a gene does not belong to any gene functional group, the entry is NA.

Value

A \(p \times p\) matrix.

See also

Author

Monika Jelizarow and Vincent Guillemot

Examples

# require(SHIP)
data(expl)
target_matrix <- targetGpos(expl$x, expl$genegroups)
which(target_matrix[upper.tri(target_matrix)] != 0) # not many non zero coefficients !
#>    [1]    1    2    3    7    8    9   16   17   18   20   37   38   39   41
#>   [15]   43   53   64   75   78   79   80   81   83   85   88  100  103  104
#>   [29]  106  107  108  110  112  115  119  129  132  133  135  137  138  139
#>   [43]  141  143  146  150  152  154  155  156  158  160  163  167  169  171
#>   [57]  172  173  174  176  178  181  185  187  189  190  211  212  213  215
#>   [71]  217  220  224  226  228  229  230  232  233  234  236  238  241  245
#>   [85]  247  249  250  251  253  277  278  279  281  283  286  290  292  294
#>   [99]  295  296  298  299  301  302  303  305  307  310  314  316  318  319
#>  [113]  320  322  323  325  326  327  328  330  332  335  339  341  343  344
#>  [127]  345  347  348  350  351  379  380  381  383  385  388  392  394  396
#>  [141]  397  398  400  401  403  404  405  407  408  409  411  413  416  420
#>  [155]  422  424  425  426  428  429  431  432  433  435  537  540  541  543
#>  [169]  545  570  573  574  576  578  595  631  632  633  637  639  640  642
#>  [183]  643  645  646  647  649  650  652  655  664  665  667  668  669  671
#>  [197]  673  676  680  682  684  685  686  688  689  691  692  693  695  696
#>  [211]  703  704  705  706  708  710  713  717  719  721  722  723  725  726
#>  [225]  728  729  730  732  733  741  788  791  931  947  948  949  951  953
#>  [239]  956  960  962  964  965  966  968  969  971  972  973  975  976  984
#>  [253]  985 1090 1093 1094 1096 1098 1115 1116 1118 1137 1140 1141 1143 1145
#>  [267] 1162 1163 1165 1176 1185 1188 1189 1191 1193 1210 1211 1213 1224 1225
#>  [281] 1234 1237 1238 1240 1242 1259 1260 1262 1273 1274 1275 1284 1287 1288
#>  [295] 1290 1292 1309 1310 1312 1323 1324 1325 1326 1331 1333 1335 1338 1339
#>  [309] 1341 1342 1343 1346 1351 1360 1361 1363 1374 1375 1376 1377 1378 1379
#>  [323] 1380 1381 1383 1385 1388 1392 1394 1396 1397 1398 1400 1401 1403 1404
#>  [337] 1405 1407 1408 1416 1417 1423 1486 1487 1488 1490 1492 1495 1499 1501
#>  [351] 1503 1504 1505 1507 1508 1510 1511 1512 1514 1515 1523 1524 1530 1539
#>  [365] 1552 1553 1555 1557 1574 1575 1577 1588 1589 1590 1591 1592 1593 1597
#>  [379] 1598 1599 1601 1603 1605 1606 1610 1612 1614 1615 1616 1618 1619 1621
#>  [393] 1622 1623 1625 1626 1630 1633 1634 1635 1641 1646 1648 1649 1650 1652
#>  [407] 1662 1665 1666 1668 1670 1687 1688 1690 1701 1702 1703 1704 1705 1710
#>  [421] 1718 1720 1721 1723 1725 1726 1728 1730 1731 1736 1745 1746 1748 1750
#>  [435] 1760 1761 1762 1763 1764 1767 1769 1771 1772 1773 1775 1777 1780 1784
#>  [449] 1786 1788 1789 1790 1792 1793 1795 1796 1797 1799 1800 1808 1809 1815
#>  [463] 1824 1826 1828 2040 2081 2082 2083 2085 2087 2090 2094 2096 2098 2099
#>  [477] 2100 2102 2103 2105 2106 2107 2109 2110 2117 2118 2119 2125 2132 2133
#>  [491] 2134 2136 2138 2140 2141 2146 2147 2148 2150 2152 2155 2159 2160 2161
#>  [505] 2162 2164 2165 2168 2170 2171 2172 2174 2175 2180 2182 2183 2184 2190
#>  [519] 2194 2195 2196 2197 2198 2199 2201 2202 2203 2205 2206 2211 2220 2223
#>  [533] 2224 2226 2228 2245 2246 2248 2259 2260 2261 2262 2263 2264 2268 2269
#>  [547] 2271 2277 2278 2416 2417 2418 2420 2422 2425 2429 2431 2433 2434 2435
#>  [561] 2437 2438 2440 2441 2442 2444 2445 2452 2453 2454 2460 2468 2469 2471
#>  [575] 2473 2476 2481 2482 2483 2557 2558 2559 2561 2563 2566 2568 2570 2571
#>  [589] 2572 2573 2574 2575 2576 2578 2579 2581 2582 2583 2585 2586 2590 2591
#>  [603] 2593 2594 2595 2601 2605 2606 2607 2608 2609 2610 2612 2613 2614 2616
#>  [617] 2617 2622 2623 2624 2627 2629 2630 2631 2633 2635 2637 2638 2642 2644
#>  [631] 2646 2647 2648 2650 2651 2653 2654 2655 2657 2658 2665 2666 2667 2673
#>  [645] 2678 2681 2682 2684 2686 2689 2694 2696 2699 2701 2776 2777 2778 2780
#>  [659] 2782 2785 2789 2791 2793 2794 2795 2797 2798 2800 2801 2802 2804 2805
#>  [673] 2813 2814 2820 2829 2831 2833 2835 2836 2841 2842 2846 2848 2849 2935
#>  [687] 2938 2939 2941 2942 2943 2945 2960 2963 2975 2976 2977 2979 2986 2994
#>  [701] 2999 3004 3005 3006 3008 3010 3013 3017 3019 3021 3022 3023 3025 3026
#>  [715] 3028 3029 3030 3032 3033 3040 3041 3042 3048 3056 3057 3059 3061 3063
#>  [729] 3064 3069 3070 3071 3074 3076 3077 3079 3241 3242 3243 3245 3247 3250
#>  [743] 3254 3256 3258 3259 3260 3262 3263 3265 3266 3267 3269 3270 3278 3279
#>  [757] 3285 3294 3296 3298 3301 3306 3307 3311 3313 3314 3316 3319 3322 3323
#>  [771] 3324 3326 3328 3331 3335 3337 3339 3340 3341 3343 3344 3346 3347 3348
#>  [785] 3350 3351 3359 3360 3366 3375 3377 3379 3381 3382 3387 3388 3392 3394
#>  [799] 3395 3397 3400 3403 3495 3498 3499 3501 3503 3520 3521 3523 3534 3535
#>  [813] 3536 3537 3538 3539 3543 3545 3546 3554 3564 3664 3667 3668 3670 3672
#>  [827] 3689 3690 3692 3703 3704 3705 3706 3707 3708 3712 3714 3715 3722 3723
#>  [841] 3728 3733 3740 3742 3743 3744 3746 3748 3751 3755 3757 3759 3760 3761
#>  [855] 3763 3764 3766 3767 3768 3770 3771 3779 3780 3786 3795 3797 3799 3801
#>  [869] 3802 3807 3808 3812 3814 3815 3817 3820 3823 3824 3829 3830 3831 3833
#>  [883] 3835 3838 3842 3844 3846 3847 3848 3850 3851 3853 3854 3855 3857 3858
#>  [897] 3865 3866 3867 3873 3881 3882 3884 3886 3888 3889 3894 3895 3899 3901
#>  [911] 3902 3904 3907 3910 3911 3916 3925 3928 3929 3931 3933 3950 3951 3953
#>  [925] 3964 3965 3966 3967 3968 3973 3975 3983 3984 3989 3994 4001 4003 4006
#>  [939] 4007 4008 4010 4012 4015 4019 4021 4023 4024 4025 4027 4028 4030 4031
#>  [953] 4032 4034 4035 4042 4043 4044 4050 4059 4061 4063 4065 4066 4071 4072
#>  [967] 4076 4078 4079 4081 4084 4087 4088 4093 4094 4380 4383 4384 4386 4388
#>  [981] 4405 4406 4419 4421 4422 4428 4430 4456 4458 4461 4496 4559 4569 4572
#>  [995] 4573 4575 4577 4594 4595 4608 4609 4610 4611 4617 4620 4627 4633 4634
#> [1009] 4638 4645 4647 4650 4689 4754 4755 4756 4758 4760 4763 4767 4769 4771
#> [1023] 4772 4773 4775 4776 4778 4779 4780 4782 4783 4791 4792 4798 4807 4809
#> [1037] 4811 4813 4814 4819 4820 4824 4826 4827 4829 4832 4835 4836 4841 4842
#> [1051] 4844 4860 4863 4864 4866 4868 4885 4886 4888 4899 4900 4901 4902 4903
#> [1065] 4908 4909 4910 4911 4919 4924 4929 4936 4938 4941 4946 4948