Signed-correlation knowledge-based target.

The \(p \times p\) target Gstar is computed from the \(n \times p\) data matrix. It it a modified version of target G. In particular, it involves two parameters for the correlation (a positive and a negative one) instead of the single parameter \(\bar{r}\) in order to account for negatively correlated genes within the same pathway.

Usage

targetGstar(x, genegroups)

Arguments

x

A \(n \times p\) data matrix.

genegroups

A list of genes obtained using the database KEGG, where each entry itself is a list of pathway names this genes belongs to. If a gene does not belong to any gene functional group, the entry is NA.

Value

A \(p \times p\) matrix.

Details

The target uses separate mean correlations for positive and negative linked pairs: $$T_{ij} = \begin{cases} \bar{r}_+\sqrt{s_{ii}s_{jj}} & \text{if } r_{ij} > 0 \\ \bar{r}_-\sqrt{s_{ii}s_{jj}} & \text{if } r_{ij} < 0 \end{cases}$$ Here \(\bar{r}_+\) and \(\bar{r}_-\) are the mean positive and negative linked correlations.

See also

Author

Monika Jelizarow and Vincent Guillemot

Examples

# require(SHIP)
data(expl)
target_matrix <- targetGstar(expl$x, expl$genegroups)
which(target_matrix[upper.tri(target_matrix)] != 0) # not many non zero coefficients !
#>    [1]    1    2    3    7    8    9   16   17   18   20   29   30   31   33
#>   [15]   35   37   38   39   41   43   45   53   56   57   58   60   62   64
#>   [29]   65   67   68   69   71   73   75   76   78   79   80   81   83   85
#>   [43]   87   88   90   91   92   93   94   96   98  100  101  103  104  105
#>   [57]  106  107  108  110  112  114  115  117  118  119  120  121  122  123
#>   [71]  125  127  129  130  132  133  134  135  136  137  138  139  141  143
#>   [85]  145  146  148  149  150  151  152  153  154  155  156  158  160  162
#>   [99]  163  165  166  167  168  169  170  171  172  173  174  176  178  180
#>  [113]  181  183  184  185  186  187  188  189  190  211  212  213  215  217
#>  [127]  219  220  222  223  224  225  226  227  228  229  230  232  233  234
#>  [141]  236  238  240  241  243  244  245  246  247  248  249  250  251  253
#>  [155]  277  278  279  281  283  285  286  288  289  290  291  292  293  294
#>  [169]  295  296  298  299  301  302  303  305  307  309  310  312  313  314
#>  [183]  315  316  317  318  319  320  322  323  325  326  327  328  330  332
#>  [197]  334  335  337  338  339  340  341  342  343  344  345  347  348  350
#>  [211]  351  379  380  381  383  385  387  388  390  391  392  393  394  395
#>  [225]  396  397  398  400  401  403  404  405  407  408  409  411  413  415
#>  [239]  416  418  419  420  421  422  423  424  425  426  428  429  431  432
#>  [253]  433  435  529  530  531  533  535  537  538  540  541  542  543  544
#>  [267]  545  546  547  548  550  551  553  554  555  557  558  562  563  564
#>  [281]  566  568  570  571  573  574  575  576  577  578  579  580  581  583
#>  [295]  584  586  587  588  590  591  595  631  632  633  635  637  639  640
#>  [309]  642  643  644  645  646  647  648  649  650  652  653  655  656  657
#>  [323]  659  660  664  665  667  668  669  671  673  675  676  678  679  680
#>  [337]  681  682  683  684  685  686  688  689  691  692  693  695  696  700
#>  [351]  701  703  704  705  706  708  710  712  713  715  716  717  718  719
#>  [365]  720  721  722  723  725  726  728  729  730  732  733  737  738  740
#>  [379]  741  788  791  931  947  948  949  951  953  955  956  958  959  960
#>  [393]  961  962  963  964  965  966  968  969  971  972  973  975  976  980
#>  [407]  981  983  984  985 1082 1083 1084 1086 1088 1090 1091 1093 1094 1095
#>  [421] 1096 1097 1098 1099 1100 1101 1103 1104 1106 1107 1108 1110 1111 1115
#>  [435] 1116 1118 1119 1120 1126 1129 1130 1131 1133 1135 1137 1138 1140 1141
#>  [449] 1142 1143 1144 1145 1146 1147 1148 1150 1151 1153 1154 1155 1157 1158
#>  [463] 1162 1163 1165 1166 1167 1173 1176 1177 1178 1179 1181 1183 1185 1186
#>  [477] 1188 1189 1190 1191 1192 1193 1194 1195 1196 1198 1199 1201 1202 1203
#>  [491] 1205 1206 1210 1211 1213 1214 1215 1221 1224 1225 1226 1227 1228 1230
#>  [505] 1232 1234 1235 1237 1238 1239 1240 1241 1242 1243 1244 1245 1247 1248
#>  [519] 1250 1251 1252 1254 1255 1259 1260 1262 1263 1264 1270 1273 1274 1275
#>  [533] 1276 1277 1278 1280 1282 1284 1285 1287 1288 1289 1290 1291 1292 1293
#>  [547] 1294 1295 1297 1298 1300 1301 1302 1304 1305 1309 1310 1312 1313 1314
#>  [561] 1320 1323 1324 1325 1326 1327 1328 1329 1331 1333 1335 1336 1338 1339
#>  [575] 1340 1341 1342 1343 1344 1345 1346 1348 1349 1351 1352 1353 1355 1356
#>  [589] 1360 1361 1363 1364 1365 1371 1374 1375 1376 1377 1378 1379 1380 1381
#>  [603] 1383 1385 1387 1388 1390 1391 1392 1393 1394 1395 1396 1397 1398 1400
#>  [617] 1401 1403 1404 1405 1407 1408 1412 1413 1415 1416 1417 1423 1426 1427
#>  [631] 1428 1429 1430 1431 1486 1487 1488 1490 1492 1494 1495 1497 1498 1499
#>  [645] 1500 1501 1502 1503 1504 1505 1507 1508 1510 1511 1512 1514 1515 1519
#>  [659] 1520 1522 1523 1524 1530 1533 1534 1535 1536 1537 1538 1539 1541 1542
#>  [673] 1543 1545 1547 1549 1550 1552 1553 1554 1555 1556 1557 1558 1559 1560
#>  [687] 1562 1563 1565 1566 1567 1569 1570 1574 1575 1577 1578 1579 1585 1588
#>  [701] 1589 1590 1591 1592 1593 1594 1596 1597 1598 1599 1601 1603 1605 1606
#>  [715] 1608 1609 1610 1611 1612 1613 1614 1615 1616 1618 1619 1621 1622 1623
#>  [729] 1625 1626 1630 1631 1633 1634 1635 1641 1644 1645 1646 1647 1648 1649
#>  [743] 1650 1652 1653 1654 1655 1656 1658 1660 1662 1663 1665 1666 1667 1668
#>  [757] 1669 1670 1671 1672 1673 1675 1676 1678 1679 1680 1682 1683 1687 1688
#>  [771] 1690 1691 1692 1698 1701 1702 1703 1704 1705 1706 1707 1709 1710 1711
#>  [785] 1712 1713 1714 1716 1718 1720 1721 1723 1724 1725 1726 1727 1728 1729
#>  [799] 1730 1731 1733 1734 1736 1737 1738 1740 1741 1745 1746 1748 1749 1750
#>  [813] 1756 1759 1760 1761 1762 1763 1764 1765 1767 1768 1769 1770 1771 1772
#>  [827] 1773 1775 1777 1779 1780 1782 1783 1784 1785 1786 1787 1788 1789 1790
#>  [841] 1792 1793 1795 1796 1797 1799 1800 1804 1805 1807 1808 1809 1815 1818
#>  [855] 1819 1820 1821 1822 1823 1824 1826 1827 1828 1829 1830 1993 2040 2081
#>  [869] 2082 2083 2085 2087 2089 2090 2092 2093 2094 2095 2096 2097 2098 2099
#>  [883] 2100 2102 2103 2105 2106 2107 2109 2110 2114 2115 2117 2118 2119 2125
#>  [897] 2128 2129 2130 2131 2132 2133 2134 2136 2137 2138 2139 2140 2141 2146
#>  [911] 2147 2148 2150 2152 2154 2155 2157 2158 2159 2160 2161 2162 2163 2164
#>  [925] 2165 2167 2168 2170 2171 2172 2174 2175 2179 2180 2182 2183 2184 2190
#>  [939] 2193 2194 2195 2196 2197 2198 2199 2201 2202 2203 2204 2205 2206 2211
#>  [953] 2212 2213 2214 2216 2218 2220 2221 2223 2224 2225 2226 2227 2228 2229
#>  [967] 2230 2231 2233 2234 2236 2237 2238 2240 2241 2245 2246 2248 2249 2250
#>  [981] 2256 2259 2260 2261 2262 2263 2264 2265 2267 2268 2269 2270 2271 2272
#>  [995] 2277 2278 2416 2417 2418 2420 2422 2424 2425 2427 2428 2429 2430 2431
#> [1009] 2432 2433 2434 2435 2437 2438 2440 2441 2442 2444 2445 2449 2450 2452
#> [1023] 2453 2454 2460 2463 2464 2465 2466 2467 2468 2469 2471 2472 2473 2474
#> [1037] 2475 2476 2481 2482 2483 2557 2558 2559 2561 2563 2565 2566 2568 2569
#> [1051] 2570 2571 2572 2573 2574 2575 2576 2578 2579 2581 2582 2583 2585 2586
#> [1065] 2590 2591 2593 2594 2595 2601 2604 2605 2606 2607 2608 2609 2610 2612
#> [1079] 2613 2614 2615 2616 2617 2622 2623 2624 2627 2629 2630 2631 2633 2635
#> [1093] 2637 2638 2640 2641 2642 2643 2644 2645 2646 2647 2648 2650 2651 2653
#> [1107] 2654 2655 2657 2658 2662 2663 2665 2666 2667 2673 2676 2677 2678 2679
#> [1121] 2680 2681 2682 2684 2685 2686 2687 2688 2689 2694 2695 2696 2699 2701
#> [1135] 2776 2777 2778 2780 2782 2784 2785 2787 2788 2789 2790 2791 2792 2793
#> [1149] 2794 2795 2797 2798 2800 2801 2802 2804 2805 2809 2810 2812 2813 2814
#> [1163] 2820 2823 2824 2825 2826 2827 2828 2829 2831 2832 2833 2834 2835 2836
#> [1177] 2841 2842 2843 2846 2848 2849 2927 2928 2929 2931 2933 2935 2936 2938
#> [1191] 2939 2940 2941 2942 2943 2944 2945 2946 2948 2949 2951 2952 2953 2955
#> [1205] 2956 2960 2961 2963 2964 2965 2971 2974 2975 2976 2977 2978 2979 2980
#> [1219] 2982 2983 2984 2985 2986 2987 2992 2993 2994 2997 2999 3000 3002 3004
#> [1233] 3005 3006 3008 3010 3012 3013 3015 3016 3017 3018 3019 3020 3021 3022
#> [1247] 3023 3025 3026 3028 3029 3030 3032 3033 3037 3038 3040 3041 3042 3048
#> [1261] 3051 3052 3053 3054 3055 3056 3057 3059 3060 3061 3062 3063 3064 3069
#> [1275] 3070 3071 3074 3076 3077 3079 3081 3241 3242 3243 3245 3247 3249 3250
#> [1289] 3252 3253 3254 3255 3256 3257 3258 3259 3260 3262 3263 3265 3266 3267
#> [1303] 3269 3270 3274 3275 3277 3278 3279 3285 3288 3289 3290 3291 3292 3293
#> [1317] 3294 3296 3297 3298 3299 3300 3301 3306 3307 3308 3311 3313 3314 3316
#> [1331] 3318 3319 3322 3323 3324 3326 3328 3330 3331 3333 3334 3335 3336 3337
#> [1345] 3338 3339 3340 3341 3343 3344 3346 3347 3348 3350 3351 3355 3356 3358
#> [1359] 3359 3360 3366 3369 3370 3371 3372 3373 3374 3375 3377 3378 3379 3380
#> [1373] 3381 3382 3387 3388 3389 3392 3394 3395 3397 3399 3400 3403 3475 3487
#> [1387] 3488 3489 3491 3493 3495 3496 3498 3499 3500 3501 3502 3503 3504 3505
#> [1401] 3506 3508 3509 3511 3512 3513 3515 3516 3520 3521 3523 3524 3525 3531
#> [1415] 3534 3535 3536 3537 3538 3539 3540 3542 3543 3544 3545 3546 3547 3552
#> [1429] 3553 3554 3557 3559 3560 3562 3564 3565 3568 3569 3633 3656 3657 3658
#> [1443] 3660 3662 3664 3665 3667 3668 3669 3670 3671 3672 3673 3674 3675 3677
#> [1457] 3678 3680 3681 3682 3684 3685 3689 3690 3692 3693 3694 3700 3703 3704
#> [1471] 3705 3706 3707 3708 3709 3711 3712 3713 3714 3715 3716 3721 3722 3723
#> [1485] 3726 3728 3729 3731 3733 3734 3737 3738 3740 3742 3743 3744 3746 3748
#> [1499] 3750 3751 3753 3754 3755 3756 3757 3758 3759 3760 3761 3763 3764 3766
#> [1513] 3767 3768 3770 3771 3775 3776 3778 3779 3780 3786 3789 3790 3791 3792
#> [1527] 3793 3794 3795 3797 3798 3799 3800 3801 3802 3807 3808 3809 3812 3814
#> [1541] 3815 3817 3819 3820 3823 3824 3826 3828 3829 3830 3831 3833 3835 3837
#> [1555] 3838 3840 3841 3842 3843 3844 3845 3846 3847 3848 3850 3851 3853 3854
#> [1569] 3855 3857 3858 3862 3863 3865 3866 3867 3873 3876 3877 3878 3879 3880
#> [1583] 3881 3882 3884 3885 3886 3887 3888 3889 3894 3895 3896 3899 3901 3902
#> [1597] 3904 3906 3907 3910 3911 3913 3915 3916 3917 3918 3919 3921 3923 3925
#> [1611] 3926 3928 3929 3930 3931 3932 3933 3934 3935 3936 3938 3939 3941 3942
#> [1625] 3943 3945 3946 3950 3951 3953 3954 3955 3961 3964 3965 3966 3967 3968
#> [1639] 3969 3970 3972 3973 3974 3975 3976 3977 3982 3983 3984 3987 3989 3990
#> [1653] 3992 3994 3995 3998 3999 4001 4003 4004 4005 4006 4007 4008 4010 4012
#> [1667] 4014 4015 4017 4018 4019 4020 4021 4022 4023 4024 4025 4027 4028 4030
#> [1681] 4031 4032 4034 4035 4039 4040 4042 4043 4044 4050 4053 4054 4055 4056
#> [1695] 4057 4058 4059 4061 4062 4063 4064 4065 4066 4071 4072 4073 4076 4078
#> [1709] 4079 4081 4083 4084 4087 4088 4090 4092 4093 4094 4095 4372 4373 4374
#> [1723] 4376 4378 4380 4381 4383 4384 4385 4386 4387 4388 4389 4390 4391 4393
#> [1737] 4394 4396 4397 4398 4400 4401 4405 4406 4408 4409 4410 4416 4419 4420
#> [1751] 4421 4422 4423 4424 4425 4427 4428 4429 4430 4431 4432 4437 4438 4439
#> [1765] 4442 4444 4445 4447 4449 4450 4453 4454 4456 4458 4459 4460 4461 4462
#> [1779] 4496 4559 4561 4562 4563 4565 4567 4569 4570 4572 4573 4574 4575 4576
#> [1793] 4577 4578 4579 4580 4582 4583 4585 4586 4587 4589 4590 4594 4595 4597
#> [1807] 4598 4599 4605 4608 4609 4610 4611 4612 4613 4614 4616 4617 4618 4619
#> [1821] 4620 4621 4626 4627 4628 4631 4633 4634 4636 4638 4639 4642 4643 4645
#> [1835] 4647 4648 4649 4650 4651 4655 4689 4754 4755 4756 4758 4760 4762 4763
#> [1849] 4765 4766 4767 4768 4769 4770 4771 4772 4773 4775 4776 4778 4779 4780
#> [1863] 4782 4783 4787 4788 4790 4791 4792 4798 4801 4802 4803 4804 4805 4806
#> [1877] 4807 4809 4810 4811 4812 4813 4814 4819 4820 4821 4824 4826 4827 4829
#> [1891] 4831 4832 4835 4836 4838 4840 4841 4842 4843 4844 4848 4850 4852 4853
#> [1905] 4854 4856 4858 4860 4861 4863 4864 4865 4866 4867 4868 4869 4870 4871
#> [1919] 4873 4874 4876 4877 4878 4880 4881 4885 4886 4888 4889 4890 4896 4899
#> [1933] 4900 4901 4902 4903 4904 4905 4907 4908 4909 4910 4911 4912 4917 4918
#> [1947] 4919 4922 4924 4925 4927 4929 4930 4933 4934 4936 4938 4939 4940 4941
#> [1961] 4942 4946 4948 4950